Adds a rudimentary navigation menu tab as well as four scenes.
The menu tab component was found in an online library.
It fits to the platform in that for android, the menu tab will
be on top, while for iOS it will be at the bottom.
The menu tab can be found and modified inside the MenuTab.js file.
The scenes are essentially the different pages of the app. They are
found inside the scenes folder.