Add Icons to Menu Tab (#15) (#18)
* Add a rudimentary event element for schedule view.

Changes are in ScheduleScene.js. Dropdown expansion has
yet to be implemented. function _onPressExpand() has no utility.

* Add trivial dropdown functionality to schedule menu.

Menu is trivial and barebones. Still needs ListView
implemented so as to allow dynamic schedule generation
based on database.

* Add relevant icon to navigation bar

Icons are different based on what platform
app is run on. There is a commented out line
to give the intro tab an icon when that is added.
New library to manage icons is also added.

* Merge fully develop and branch

Also modified RGB color values so as
to reference shared color style Jack made.

* Make coding style changes